src/requirements.txt
requests
discord.py-self[voice]==2.0
aiohttp==3.8.5
tls-client
